# api/openapi.yaml
openapi: 3.0.3
info:
  title: IGNAI Hackathon Platform API
  version: 0.4.0
  description: 'Spring Boot / MyBatis 版本接口规划。MVP 无独立评委角色，无视频上传； 后台管理入口规划为 /root/，root
    初始化和认证使用 /api/root/*。

    '
servers:
- url: http://localhost:8080/api
  description: Local Spring Boot API
...
paths:
...
  /public/projects:
    get:
      tags:
      - Project
      summary: 公开作品列表
      description: |
        ## 前台作品展示列表
        返回已发布或已提交的作品，支持多维度筛选和排序。
        - **无需认证**：访客浏览作品墙
        - **筛选**：`status`（published/submitted）、`trackId`（赛道）、`sort`（latest/votes/rating/featured）
        - **用途**：前台作品展示页、排行榜
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - published
          - submitted
      - name: trackId
        in: query
        schema:
          type: integer
      - name: sort
        in: query
        schema:
          type: string
          enum:
          - latest
          - votes
          - rating
          - featured
      - name: pageSize
        in: query
        schema:
          type: integer
      responses:
        '200':
          description: 作品列表
          content:
            application/json:
              schema:
                type: object
      operationId: public_projects_get
  /public/projects/{projectId}:
    get:
      tags:
      - Project
      summary: 公开作品详情
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 作品详情
          content:
            application/json:
              schema:
                type: object
      operationId: public_projects_projectId_get
...
  /projects:
    post:
      tags:
      - Project
      summary: 创建或保存作品
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpsertRequest'
      responses:
        '200':
          description: 作品信息
          content:
            application/json:
              schema:
                type: object
      operationId: projects_post
  /projects/my:
    get:
      tags:
      - Project
      summary: 我的作品
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 当前用户作品
          content:
            application/json:
              schema:
                type: object
      operationId: projects_my_get
  /projects/{projectId}:
    put:
      tags:
      - Project
      summary: 更新作品
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
      operationId: projects_projectId_put
  /projects/{projectId}/files:
    post:
      tags:
      - Project
      summary: 上传附件
      description: MVP 不允许上传视频文件。
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema:
                type: object
        '400':
          description: 文件类型不允许
          content:
            application/json:
              schema:
                type: object
      operationId: projects_projectId_files_post
  /projects/{projectId}/submit:
    post:
      tags:
      - Project
      summary: 正式提交作品
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 提交成功
          content:
            application/json:
              schema:
                type: object
      operationId: projects_projectId_submit_post
  /admin/projects:
    get:
      tags:
      - Project
      summary: 后台项目看板
      security:
      - bearerAuth: []
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - draft
          - submitted
          - locked
          - published
          - archived
      - name: trackId
        in: query
        schema:
          type: integer
      - name: sort
        in: query
        schema:
          type: string
          enum:
          - latest
          - votes
          - rating
      - name: keyword
        in: query
        schema:
          type: string
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                type: object
      operationId: admin_projects_get
  /admin/projects/{projectId}:
    get:
      tags:
      - Project
      summary: 后台项目详情
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 项目详情
          content:
            application/json:
              schema:
                type: object
      operationId: admin_projects_projectId_get
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ProjectUpsertRequest:
      type: object
      required:
      - eventId
      - teamId
      - title
      properties:
        eventId:
          type: integer
        teamId:
          type: integer
        trackId:
          type: integer
        title:
          type: string
        tagline:
          type: string
        description:
          type: string
        demoUrl:
          type: string
        codeUrl:
          type: string
        payload:
          type: object
...